cron/requirements.txt
anyio==3.6.2
astroid==2.15.2
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dill==0.3.6
exceptiongroup==1.1.1
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.2.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pylint==2.17.2
pytest==7.3.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
